+++ src/ts/type.ts
@@ -4,7 +4,7 @@
  * A lightweight JavaScript library that attaches a viewer to a specific node type, allowing you to view the CSS properties, attributes, and size/position.
  * 
  * @file        type.ts
- * @version     v1.4.0
+ * @version     v1.5.0
  * @author      Bunoon
  * @license     MIT License
  * @copyright   Bunoon 2024
@@ -17,13 +17,18 @@ export type Position = {
 };
 
 export type Configuration = {
+	dialogDisplayDelay?: number;
+    searchDelayDelay?: number;
+    text?: ConfigurationText;
+};
+
+export type ConfigurationText = {
 	cssText?: string;
     attributesText?: string;
     sizeText?: string;
     classesText?: string;
     noAttributesAvailableText?: string;
     closeText?: string;
-    dialogDisplayDelay?: number;
     copyText?: string;
     copySymbolText?: string;
     pasteText?: string;
@@ -31,7 +36,11 @@ export type Configuration = {
     removeText?: string;
     removeSymbolText?: string;
     noClassesAvailableText?: string;
-};
+    searchPropertiesPlaceHolderText?: string;
+    clearText?: string;
+    clearSymbolText?: string;
+    noPropertiesFoundForSearchText?: string;
+}
 
 export type Options = {
     nodeType: string[] | string;